cron-jobs/requirements.txt
feedparser
sqlalchemy
psycopg2
newspaper3k